requirements.txt
PyQt5>=5.15.11
pyqtgraph>=0.13.7
pyserial>=3.5
numpy>=2.0,<3
PyYAML>=6.0
iso8601>=1.1.0
future>=0.18.3
altgraph>=0.17.3

# Windows-only helpers
pefile>=2023.2.7; platform_system == "Windows"
pywin32-ctypes>=0.2.2; platform_system == "Windows"
